client-ui/requirements.txt
phe 
Flask-API 
flask_cors 
requests
pyopenssl
pymongo
dnspython
pycryptodome
cryptocode